Refactor GUI into modular package layout
